requirements.txt
torch>=2.0.0
numpy>=1.24.0
biopython>=1.81
requests>=2.31.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
einops>=0.6.1
